feat(prod): un BDT n'est sur le planning que s'il y a ete pose ; double-clic = recevoir
« Quand les BDT sont a programmer ils ne doivent pas apparaitre dans le
planning : ils sont programmes quand on les passe de la goulotte au planning. »
(11/09/2026)

POURQUOI DES BDT S'ETAIENT « PLACES TOUT SEULS » : l'appartenance au Gantt se
lisait sur le process, avec repli sur le LIBELLE de l'operation, et un BDT sans
date etait rattache au jour courant a 6 h. Depuis le 10/09 plus aucun BDT n'est
filtre cote serveur : 3 des 4 BDT de l'affaire 0001 etaient A LA FOIS dans la
goulotte et sur le planning. En plus, la cascade creait les BDT « programmes ».
-> enGoulotte(b) : statut a_programmer, sans process OU SANS JOUR => goulotte.
   Gantt, goulotte, stats et reception utilisent ce seul critere. Tous les
   createurs de BDT ecrivent a_programmer.

PROGRAMMER AU QUART D'HEURE : bons_de_travail.debut etait un ENTIER, tout depot
a h15/h30/h45 echouait (« Affectation echouee »). Migration 005 (VM : seule) +
cloud-4 ; sans elle le serveur arrondit a l'heure pleine au lieu d'echouer.
/affecter : jour toujours ecrit, 409 si recu/solde, machine suit le process.

DOUBLE-CLIC : programme -> reception (matricule + PIN) ; recu -> sortie matiere.
/recu refuse (409) un BDT non programme, deja recu ou solde, et note l'heure de
PARIS (serveur en UTC : temps reel majore de 2 h en ete).
DEPROGRAMMER : retour complet en goulotte ; 409 si recu/solde.
SEPARER depuis la goulotte (ciseaux). Les morceaux naissaient programmes et
tombaient sur le planning ; l'original etait rogne AVANT leur creation (echecs
avales) ; une 2e separation reutilisait -M2 (cle primaire). Morceaux en goulotte,
crees d'abord avec retour arriere (supprimerBDTRow), suffixe depuis la racine.
PORTE DES BST (gamme.ts) : un BDT « recu » (demarre) n'est plus « solde » ; tous
les morceaux d'une etape doivent etre soldes. /production/gantt-bdt (ancien
planning, regles differentes) redirige vers /production/service?focusBdt=.

Verifie sur Docker, jeu -TEST- supprime : 7h15 -> 7,25 ; reception refusee hors
planning, acceptee au PIN a l'heure locale ; reçu non replanifiable ;
-M2/-M3/-M4 sans collision, heures conservees ; navigateur : les 4 BDT de 0001
en goulotte, double-clic -> recuModal puis sortieModal. gamme.ts : 9/9.

+++ docker/db/seed/hygiene.sql
@@ -26,7 +26,7 @@
 -- Types déduits des valeurs réelles du cloud ; cout_reel et avancement sont typés
 -- numériques d'après leur usage dans le code (Number(...) et pourcentage arrondi).
 alter table public.bons_de_travail add column if not exists client_nom          text;
-alter table public.bons_de_travail add column if not exists debut               integer;
+alter table public.bons_de_travail add column if not exists debut               numeric;
 alter table public.bons_de_travail add column if not exists priorite            text;
 alter table public.bons_de_travail add column if not exists baseline            text;
 alter table public.bons_de_travail add column if not exists pv_requis           boolean;
